Criando pasta utils para reutilização de funções
